Move load-or-simulate function to new workflows.scripting module.  Make toast_so_sim and toast_so_map as consistent as possible.
