+++ sotodlib/toast/scripts/so_sim.py
@@ -49,35 +49,6 @@ import pixell.fft
 pixell.fft.engine = "fftw"
 
 
-def load_or_simulate_observing(job, otherargs, runargs, comm):
-    log = toast.utils.Logger.get()
-    job_ops = job.operators
-
-    if job_ops.sim_ground.enabled:
-        data = wrk.simulate_observing(job, otherargs, runargs, comm)
-        wrk.simple_noise_models(job, otherargs, runargs, data)
-    else:
-        group_size = wrk.reduction_group_size(job, runargs, comm)
-        toast_comm = toast.Comm(world=comm, groupsize=group_size)
-        data = toast.Data(comm=toast_comm)
-        # Load data from all formats
-        wrk.load_data_hdf5(job, otherargs, runargs, data)
-        wrk.load_data_books(job, otherargs, runargs, data)
-        wrk.load_data_context(job, otherargs, runargs, data)
-        wrk.act_responsivity_sign(job, otherargs, runargs, data)
-        wrk.create_az_intervals(job, otherargs, runargs, data)
-        wrk.apply_readout_filter(job, otherargs, runargs, data)
-        # optionally zero out
-        if otherargs.zero_loaded_data:
-            toast.ops.Reset(detdata=[defaults.det_data])
-        # Append a weather model
-        wrk.append_weather_model(job, otherargs, runargs, data)
-        # Before running simulations based on real data, we need
-        # a starting noise estimate
-        wrk.diff_noise_estimation(job, otherargs, runargs, data)
-    wrk.select_pointing(job, otherargs, runargs, data)
-    return data
-
 def simulate_data(job, otherargs, runargs, data):
     log = toast.utils.Logger.get()
 
@@ -118,8 +89,8 @@ def simulate_data(job, otherargs, runargs, data):
 def reduce_data(job, otherargs, runargs, data):
     log = toast.utils.Logger.get()
 
-    wrk.simple_deglitch(job, otherargs, runargs, data)
     wrk.simple_jumpcorrect(job, otherargs, runargs, data)
+    wrk.simple_deglitch(job, otherargs, runargs, data)
 
     wrk.flag_diff_noise_outliers(job, otherargs, runargs, data)
     wrk.flag_noise_outliers(job, otherargs, runargs, data)
@@ -135,13 +106,13 @@ def reduce_data(job, otherargs, runargs, data):
     wrk.noise_estimation(job, otherargs, runargs, data)
 
     data = wrk.demodulate(job, otherargs, runargs, data)
-    
+
     wrk.processing_mask(job, otherargs, runargs, data)
     wrk.flag_sso(job, otherargs, runargs, data)
     wrk.hn_map(job, otherargs, runargs, data)
     wrk.cadence_map(job, otherargs, runargs, data)
     wrk.crosslinking_map(job, otherargs, runargs, data)
-    
+
     wrk.mapmaker_ml(job, otherargs, runargs, data)
     wrk.mapmaker(job, otherargs, runargs, data)
     wrk.mapmaker_filterbin(job, otherargs, runargs, data)
@@ -219,20 +190,8 @@ def main():
     operators = list()
     templates = list()
 
-    # Loading data from disk is disabled by default
-    wrk.setup_load_data_hdf5(operators)
-    wrk.setup_load_data_books(operators)
-    wrk.setup_load_data_context(operators)
-    wrk.setup_act_responsivity_sign(operators)
+    wrk.setup_load_or_simulate_observing(parser, operators)
 
-    # Simulated observing is enabled by default
-    wrk.setup_simulate_observing(parser, operators)
-
-    wrk.setup_pointing(operators)
-    wrk.setup_az_intervals(operators)
-    wrk.setup_readout_filter(operators)
-    wrk.setup_simple_noise_models(operators)
-    wrk.setup_weather_model(operators)
     wrk.setup_simulate_atmosphere_signal(operators)
     wrk.setup_simulate_sky_map_signal(operators)
     wrk.setup_simulate_conviqt_signal(operators)
@@ -251,24 +210,31 @@ def main():
     wrk.setup_simulate_readout_effects(operators)
     wrk.setup_save_data_hdf5(operators)
 
+    wrk.setup_simple_jumpcorrect(operators)
+    wrk.setup_simple_deglitch(operators)
+
     wrk.setup_flag_diff_noise_outliers(operators)
     wrk.setup_flag_noise_outliers(operators)
+    wrk.setup_deconvolve_detector_timeconstant(operators)
+    wrk.setup_raw_statistics(operators)
+
     wrk.setup_filter_hwpss(operators)
-    wrk.setup_demodulate(operators)
+    wrk.setup_filter_common_mode(operators)
+    wrk.setup_filter_ground(operators)
+    wrk.setup_filter_poly1d(operators)
+    wrk.setup_filter_poly2d(operators)
     wrk.setup_diff_noise_estimation(operators)
     wrk.setup_noise_estimation(operators)
+
+    wrk.setup_demodulate(operators)
+
     wrk.setup_processing_mask(operators)
     wrk.setup_flag_sso(operators)
     wrk.setup_hn_map(operators)
     wrk.setup_cadence_map(operators)
     wrk.setup_crosslinking_map(operators)
-    wrk.setup_raw_statistics(operators)
-    wrk.setup_deconvolve_detector_timeconstant(operators)
+
     wrk.setup_mapmaker_ml(operators)
-    wrk.setup_filter_ground(operators)
-    wrk.setup_filter_poly1d(operators)
-    wrk.setup_filter_poly2d(operators)
-    wrk.setup_filter_common_mode(operators)
     wrk.setup_mapmaker(operators, templates)
     wrk.setup_mapmaker_filterbin(operators)
     wrk.setup_mapmaker_madam(operators)
@@ -292,8 +258,8 @@ def main():
         log.info_rank("Dry-run complete", comm=comm)
         return
 
-    data = load_or_simulate_observing(job, otherargs, runargs, comm)
-    
+    data = wrk.load_or_simulate_observing(job, otherargs, runargs, comm)
+
     simulate_data(job, otherargs, runargs, data)
 
     if not job.operators.sim_atmosphere.cache_only:
